--- a/src/StackExchange.Redis/PhysicalBridge.cs
+++ b/src/StackExchange.Redis/PhysicalBridge.cs
@@ -836,6 +836,10 @@ namespace StackExchange.Redis
                 thread.Start(this);
 #endif
             }
+            else
+            {
+                _backlogAutoReset.Set();
+            }
         }
 
         /// <summary>
@@ -909,7 +913,7 @@ namespace StackExchange.Redis
                     // TODO: vNext handoff this backlog to another primary ("can handle everything") connection
                     // and remove any per-server commands. This means we need to track a bit of whether something
                     // was server-endpoint-specific in PrepareToPushMessageToBridge (was the server ref null or not)
-                    await ProcessBridgeBacklogAsync(); // Needs handoff
+                    await ProcessBridgeBacklogAsync().ConfigureAwait(false); // Needs handoff
                 }
             }
             catch
@@ -938,6 +942,13 @@ namespace StackExchange.Redis
             }
         }
 
+        /// <summary>
+        /// Reset event for monitoring backlog additions mid-run.
+        /// This allows us to keep the thread around for a full flush and prevent "feathering the throttle" trying
+        /// to flush it. In short, we don't start and stop so many threads with a bit of linger.
+        /// </summary>
+        private readonly AutoResetEvent _backlogAutoReset = new AutoResetEvent(false);
+
         private async Task ProcessBridgeBacklogAsync()
         {
             // Importantly: don't assume we have a physical connection here
@@ -947,6 +958,7 @@ namespace StackExchange.Redis
 #else
             LockToken token = default;
 #endif
+            _backlogAutoReset.Reset();
             try
             {
                 _backlogStatus = BacklogStatus.Starting;
@@ -985,7 +997,20 @@ namespace StackExchange.Redis
                     {
                         // Note that we're actively taking it off the queue here, not peeking
                         // If there's nothing left in queue, we're done.
-                        if (!BacklogTryDequeue(out message)) break;
+                        if (!BacklogTryDequeue(out message))
+                        {
+                            // The cost of starting a new thread is high, and we can bounce in and out of the backlog a lot.
+                            // So instead of just exiting, keep this thread waiting for 5 seconds to see if we got another backlog item.
+                            var gotMore = _backlogAutoReset.WaitOne(5000);
+                            if (gotMore)
+                            {
+                                continue;
+                            }
+                            else
+                            {
+                                break;
+                            }
+                        }
                     }
 
                     try